requirements.txt
asgiref==3.2.10
boto3==1.14.20
botocore==1.17.20
dj-database-url==0.5.0
Django==3.0.8
django-storages==1.9.1
djangorestframework==3.11.0
docutils==0.15.2
jmespath==0.10.0
Pillow==2.2.1
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
whitenoise==5.1.0
